// src/generated/models/ProjectMember.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Codebook Api
 * A simple api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RoleEnum } from './RoleEnum';
import {
    RoleEnumFromJSON,
    RoleEnumFromJSONTyped,
    RoleEnumToJSON,
} from './RoleEnum';
import type { UserSummary } from './UserSummary';
import {
    UserSummaryFromJSON,
    UserSummaryFromJSONTyped,
    UserSummaryToJSON,
} from './UserSummary';

/**
 * 
 * @export
 * @interface ProjectMember
 */
export interface ProjectMember {
    /**
     * 
     * @type {number}
     * @memberof ProjectMember
     */
    readonly id: number;
    /**
     * 
     * @type {UserSummary}
     * @memberof ProjectMember
     */
    user: UserSummary;
    /**
     * 
     * @type {RoleEnum}
     * @memberof ProjectMember
     */
    role?: RoleEnum;
    /**
     * 
     * @type {number}
     * @memberof ProjectMember
     */
    project: number;
}

/**
 * Check if a given object implements the ProjectMember interface.
 */
export function instanceOfProjectMember(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "user" in value;
    isInstance = isInstance && "project" in value;

    return isInstance;
}

export function ProjectMemberFromJSON(json: any): ProjectMember {
    return ProjectMemberFromJSONTyped(json, false);
}

export function ProjectMemberFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectMember {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'user': UserSummaryFromJSON(json['user']),
        'role': !exists(json, 'role') ? undefined : RoleEnumFromJSON(json['role']),
        'project': json['project'],
    };
}

export function ProjectMemberToJSON(value?: ProjectMember | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': UserSummaryToJSON(value.user),
        'role': RoleEnumToJSON(value.role),
        'project': value.project,
    };
}