// src/generated/models/ProjectMemberMutationRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Codebook Api
 * A simple api
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { RoleEnum } from './RoleEnum';
import {
    RoleEnumFromJSON,
    RoleEnumFromJSONTyped,
    RoleEnumToJSON,
} from './RoleEnum';

/**
 * 
 * @export
 * @interface ProjectMemberMutationRequest
 */
export interface ProjectMemberMutationRequest {
    /**
     * 
     * @type {number}
     * @memberof ProjectMemberMutationRequest
     */
    user: number;
    /**
     * 
     * @type {RoleEnum}
     * @memberof ProjectMemberMutationRequest
     */
    role?: RoleEnum;
}

/**
 * Check if a given object implements the ProjectMemberMutationRequest interface.
 */
export function instanceOfProjectMemberMutationRequest(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "user" in value;

    return isInstance;
}

export function ProjectMemberMutationRequestFromJSON(json: any): ProjectMemberMutationRequest {
    return ProjectMemberMutationRequestFromJSONTyped(json, false);
}

export function ProjectMemberMutationRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProjectMemberMutationRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'user': json['user'],
        'role': !exists(json, 'role') ? undefined : RoleEnumFromJSON(json['role']),
    };
}

export function ProjectMemberMutationRequestToJSON(value?: ProjectMemberMutationRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'user': value.user,
        'role': RoleEnumToJSON(value.role),
    };
}